==> 5-Supp_figures.R <==
library(tidyverse)
library(lubridate)
library(ggmap)
library(cowplot)
library(gridExtra)
library(ggthemes)
library(feather)
library(RColorBrewer)
library(MASS)
library(plotly)
library(gg3D)
library(marmap)

JSD <- function(p, q){
  # pfun <- approxfun(density(p))
  # p <- pfun(p)/sum(pfun(p))
  # qfun <- approxfun(density(q))
  # q <- qfun(q)/sum(qfun(q))
  m <- 0.5 * (p + q)
  JS <- 0.5 * (sum(p * log(p / m)) + sum(q * log(q / m)))
  return(JS)
}

x = ph
y = qh
JSD(x, y)
JSD<- function(x,y) sqrt(0.5 * (sum(x*log(x/((x+y)/2))) + sum(y*log(y/((x+y)/2)))))


#--------------------------------------------------------------------------------------------
# Figure S1: Distance (log transformation))

# (A)

dat <- as.data.frame(read_feather('~/Projects/Anomalous-IUU-Events-Argentina/data/Argentina_5NN_region1_2018-01-15_2018-02-15.feather'))
dat <- dat %>% 
mutate(day = day(timestamp),
       hour = hour(timestamp),
       month = month(timestamp)) %>% 
filter(distance != 0) %>% 
filter(timestamp >= "2018-01-16 00:00:00" & timestamp <= "2018-02-15 23:00:00") %>% 
group_by(timestamp, vessel_A) %>% 
summarise(distance = mean(distance),
          ln_distance = mean(log(1 + distance)))

dat2$date <- paste0(year(dat2$timestamp), "-", month(dat2$timestamp), "-", day(dat2$timestamp), "-", hour(dat2$timestamp))

cdat <- cut(dat$ln_distance, breaks = 50)

dat$breaks <- cdat 
dat$bin <- as.numeric(dat$breaks)


char <- unique(cdat)
retdat <- data.frame()
for (j in char){
  ldat <- data.frame(a = as.numeric(strsplit(gsub("\\[|\\]|\\(|\\)", "", j), ",")[[1]][1]),
                     b = as.numeric(strsplit(gsub("\\[|\\]|\\(|\\)", "", j), ",")[[1]][2]))
  retdat <- rbind(retdat, ldat)
}

outdat <- dat %>% 
  group_by(timestamp) %>% 
  mutate(nvessels = n()) %>% 
  group_by(timestamp, bin, nvessels) %>% 
  summarise(nbin_vessels = n()) %>% 
  mutate(prob = nbin_vessels/nvessels) %>% 
  ungroup()

outdat %>% 
  group_by(timestamp) %>% 
  summarise(sum = sum(prob))

outdat$day <- day(outdat$timestamp)


outdat$timestamp
as.Date.POSIXct(outdat$timestamp, c("%Y-%m-%d %h:00:00 PST"))
outdat$timestamp <- as.POSIXct(outdat$timestamp)

sb <- c(seq(5, 45, 5))
# Get log feet to dispaly on left side
ddat <- dat %>% 
  group_by(bin) %>% 
  summarise(m_dist = mean(ln_distance)) %>% 
  filter(bin %in% sb)
ddat

outdat <- filter(outdat, bin >= 5 & bin <= 45)

fig2a <- ggplot(outdat, aes(x=timestamp, y=factor(bin))) + 
  geom_tile(aes(fill = prob)) +
  theme_tufte(14) +
  annotate("text", x=as.POSIXct("2018-02-14 8:00:00"), y = 41.5, label='(A)', size = 5, color  = "white") +
  labs(x="Day in Jan-Feb", y="Binned Distance(log km)", fill="P(d)") +
  scale_x_datetime(date_breaks = "2 day",
                   date_labels = "%d",
                   expand = expand_scale(mult = c(0, 0))) +
  scale_y_discrete(breaks = c(5, 10, 15, 20, 25, 30, 35, 40, 45), 
                   labels = round(ddat$m_dist[1:9], 1), 
                   expand = c(0, 0)) +
  scale_fill_gradientn(colours=rev(brewer.pal(11, "Spectral")), na.value = 'salmon') +
  annotate("rect", xmin = as.POSIXct("2018-01-30 01:00:00"),
           xmax = as.POSIXct("2018-02-05 01:00:00"),
           ymin = 0,
           ymax = 43,
           colour="white", alpha=0.1) +
  annotate("text", x = as.POSIXct("2018-02-02 00:00:00"), y=40, label="Event \n Window", color='white') +
  theme(legend.position = 'right',
        legend.margin=margin(l = 0, unit='cm'),
        panel.border = element_rect(colour = "grey", fill=NA, size=1),
        panel.grid = element_blank(),
        panel.background=element_rect(fill="#5E4FA2", colour="#5E4FA2")
  ) +
  guides(fill = guide_colorbar(label.hjust = unit(0, 'cm'),
                               frame.colour = "black",
                               barwidth = .5,
                               barheight = 10)) +
  NULL

fig2a

# Figure 2 (B) 

dat <- as.data.frame(read_feather("~/Projects/Anomalous-IUU-Events-Argentina/data/dmat_Puerto_Madryn_region1_NN1_day_2018-01-15_2018-02-15.feather"))

dat$day <- seq(1, nrow(dat), 1)
pdat <- gather(dat, key = day, value = value)

pdat$day <- as.numeric(pdat$day) + 1
pdat$day2 <- seq(1, length(unique(pdat$day)))

fig2b <- ggplot(pdat, aes(x=day, y=day2)) + 
  theme_tufte(14) + 
  geom_tile(aes(fill=value)) +
  labs(y="Day in Jan-Feb", x="Day in Jan-Feb", fill="JS \nMetric") +
  scale_fill_gradientn(colours=rev(brewer.pal(11, "Spectral"))) +
  scale_y_continuous(trans = "reverse", breaks = c(1, 4, 9, 14, 21, 26, 31),
                     labels = c(16, 20, 25, 30, 5, 10, 15), expand=expand_scale(mult = c(0, 0))) +
  scale_x_continuous(breaks = c(1, 4, 9, 14, 21, 26, 31),
                     labels = c(16, 20, 25, 30, 5, 10, 15), expand=expand_scale(mult = c(0, 0))) +
  annotate("segment", x=-Inf, xend=Inf, y=-Inf, yend=-Inf, color = "black") +
  annotate("segment", x=-Inf, xend=-Inf, y=-Inf, yend=Inf, color = "black") +
  annotate("segment", x=Inf, xend=Inf, y=-Inf, yend=Inf, color = "black") +
  annotate("segment", x=Inf, xend=-Inf, y=Inf, yend=Inf, color = "black") +
  
  # Cluster 1
  annotate("segment", x=15, xend=15, y=0.5, yend=15, color = "black") +
  annotate("segment", x=0.5, xend=15, y=15, yend=15, color = "black") +
  #
  # # Cluster 2
  annotate("segment", x=15, xend=22, y=15, yend=15, color = "black") +
  annotate("segment", x=15, xend=15, y=15, yend=22, color = "black") +
  annotate("segment", x=22, xend=22, y=15, yend=22, color = "black") +
  annotate("segment", x=15, xend=22, y=22, yend=22, color = "black") +
  # #
  # # Cluster 3
  annotate("segment", x=22, xend=22, y=22, yend=32.5, color = "black") +
  annotate("segment", x=22, xend=32.5, y=22, yend=22, color = "black") +
  # # 
  annotate("text", x=12, y = 1.5, label='Cluster 1', size = 3) +
  annotate("text", x=19, y = 16, label='Cluster 2', size = 3) +
  annotate("text", x=30, y = 23, label='Cluster 3', size = 3) +
  annotate("text", x=31, y = 1.5, label='(B)', size = 5, color="white") +
  guides(fill = guide_colorbar(label.hjust = unit(0, 'cm'),
                               frame.colour = "black",
                               barwidth = .5,
                               barheight = 12)) +
  coord_equal() +
  NULL
fig2b

ggdraw() + draw_plot(fig2a, 0, .50, height = .50, width = 1) +
  draw_plot(fig2b, 0.01, -.25, height=1, width = 1)

ggsave("~/Projects/Anomalous-IUU-Events-Argentina/figures/figure_s1.png", width = 5, height = 8)
ggsave("~/Projects/Anomalous-IUU-Events-Argentina/figures/figure_s1.pdf", width = 5, height = 8)


#--------------------------------------------------------------------------------------------
# Figure S2 - time-series of the trailing rate of change in JS divergence as an index of behavioral change
dat <- as.data.frame(read_feather("~/Projects/Anomalous-IUU-Events-Argentina/data/dmat_Puerto_Madryn_region1_NN1_day-hour_2018-01-15_2018-02-15.feather"))

d <- as.matrix(dat)
fit <- isoMDS(d, k=2)
stress <- fit$stress

isoMDS_dat <- data.frame(x = fit$points[, 1], y = fit$points[, 2])


# Hourly cluster
#Clusters: 

cluster1 <- c(216, 0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 18, 19, 20, 21, 22, 23, 24, 25, 26, 27, 28, 29, 30, 31, 32, 33, 34, 35, 36, 37, 38, 39, 40, 41, 42, 43, 44, 45, 46, 47, 48, 49, 50, 51, 52, 53, 54, 55, 56, 57, 58, 59, 60, 61, 62, 63, 64, 65, 66, 67, 68, 69, 70, 71, 72, 73, 74, 75, 76, 77, 78, 79, 80, 81, 82, 83, 84, 85, 86, 87, 88, 89, 90, 91, 92, 93, 94, 95, 96, 97, 98, 99, 100, 101, 102, 103, 104, 105, 106, 107, 108, 109, 110, 111, 112, 113, 114, 115, 116, 117, 118, 119, 120, 121, 122, 123, 124, 125, 126, 127, 128, 129, 130, 131, 132, 133, 134, 135, 136, 137, 138, 139, 140, 141, 142, 143, 144, 145, 146, 147, 148, 149, 150, 151, 152, 153, 154, 155, 156, 157, 158, 159, 160, 161, 162, 163, 164, 165, 166, 167, 168, 169, 170, 171, 172, 173, 174, 175, 176, 177, 178, 179, 180, 181, 182, 183, 184, 185, 186, 187, 188, 189, 190, 191, 192, 193, 194, 195, 196, 197, 198, 199, 200, 201, 202, 203, 204, 205, 206, 207, 208, 209, 210, 211, 212, 213, 214, 215, 217, 218, 219, 220, 221, 222, 223, 224, 225, 226, 227, 228, 229, 230, 231, 232, 233, 234, 235, 236, 237, 238, 239, 240, 241, 242, 243, 244, 245, 246, 247, 248, 249, 250, 251, 252, 253, 254, 255, 256, 257, 258, 259, 260, 261, 262, 263, 264, 265, 266, 267, 268, 269, 270, 271, 272, 273, 274, 275, 276, 277, 278, 279, 280, 281, 282, 283, 284, 285, 286, 287, 288, 289, 290, 291, 292, 293, 294, 295, 296, 297, 298, 299, 300, 301, 302, 303, 304, 305, 306, 307, 308, 309, 310, 311, 312, 313, 314, 315, 316, 317, 318, 319, 320, 321, 322, 323, 324, 325, 326, 327, 328, 329, 330, 331, 332, 333, 334, 335, 336, 337, 338, 339, 340, 341, 342, 343, 344, 345, 346, 347, 348, 349, 350, 351, 352, 353, 354, 355, 356, 357, 358, 359, 360, 361, 362, 363, 364, 365, 366, 367, 368, 369, 370, 371, 372, 373, 374, 375, 376, 377, 379, 380, 381, 382, 383, 384, 386, 388)
cluster2 <- c(430, 378, 385, 387, 389, 390, 391, 392, 393, 394, 395, 396, 397, 398, 399, 400, 401, 402, 403, 404, 405, 406, 407, 408, 409, 410, 411, 412, 413, 414, 415, 416, 417, 418, 419, 420, 421, 422, 423, 424, 425, 426, 427, 428, 429, 431, 432, 433, 434, 435, 436, 437, 438, 439, 440, 441, 442, 443, 444, 445, 446, 447, 448, 449, 450, 451, 452, 453, 454, 455, 456, 457, 458, 459, 460, 461, 462, 463, 464, 465, 466, 467, 468, 469, 470, 471, 472, 473, 474, 475, 476, 477, 478, 479, 480, 481, 482, 483, 484, 485, 486, 488, 489, 490, 492, 494)
cluster3 <- c(572, 487, 491, 493, 495, 496, 497, 498, 499, 500, 501, 502, 503, 504, 505, 506, 507, 508, 509, 510, 511, 512, 513, 514, 515, 516, 517, 518, 519, 520, 521, 522, 523, 524, 525, 526, 527, 528, 529, 530, 531, 532, 533, 534, 535, 536, 537, 538, 539, 540, 541, 542, 543, 544, 545, 546, 547, 548, 549, 550, 551, 552, 553, 554, 555, 556, 557, 558, 559, 560, 561, 562, 563, 564, 565, 566, 567, 568, 569, 570, 571, 573, 574, 575, 576, 577, 578, 579, 580, 581, 582, 583, 584, 585, 586, 587, 588, 589, 590, 591, 592, 593, 594, 595, 596, 597, 598, 599, 600, 601, 602, 603, 604, 605, 606, 607, 608, 609, 610, 611, 612, 613, 614, 615, 616, 617, 618, 619, 620, 621, 622, 623, 624, 625, 626, 627, 628, 629, 630, 631, 632, 633, 634, 635, 636, 637, 638, 639, 640, 641, 642, 643, 644, 645, 646, 647, 648, 649, 650, 651, 652, 653, 654, 655, 656, 657, 658, 659, 660, 661, 662, 663, 664, 665, 666, 667, 668, 669, 670, 671, 672, 673, 674, 675, 676, 677, 678, 679, 680, 681, 682, 683, 684, 685, 686, 687, 688, 689, 690, 691, 692, 693, 694, 695, 696, 697, 698, 699, 700, 701, 702, 703, 704, 705, 706, 707, 708, 709, 710, 711, 712, 713, 714, 715, 716, 717, 718, 719, 720, 721, 722, 723, 724, 725, 726, 727, 728, 729, 730, 731, 732, 733, 734, 735, 736, 737, 738, 739, 740, 741, 742, 743, 744, 745, 746, 747, 748, 749, 750, 751, 752, 753, 754, 755, 756, 757, 758, 759, 760, 761, 762, 763, 764, 765, 766, 767)

#Medoids: [216, 430, 572]

# Hours
isoMDS_dat$row <- seq(0, nrow(isoMDS_dat) - 1)

# Merge data for cluster factors
clustdat <- data.frame(cluster = c(rep(1, length(cluster1)),
                                   rep(2, length(cluster2)),
                                   rep(3, length(cluster3))),
                       row = c(cluster1, cluster2, cluster3))

# Merge data
isoMDS_dat <- left_join(isoMDS_dat, clustdat, by = 'row')

# Calc distance t and t+1
# 2-axis
isoMDS_dat$dist <- sqrt( (isoMDS_dat$x1 - lead(isoMDS_dat$x1))^2 + (isoMDS_dat$x2- lead(isoMDS_dat$x2))^2 )

# Remove last obs
isoMDS_dat <- filter(isoMDS_dat, !is.na(cluster) | !is.na(dist))

event_day <- filter(isoMDS_dat, row >= 17*24 & row <= 18*24)

# Remove last obs
isoMDS_dat <- filter(isoMDS_dat, !is.na(dist))

# Calculate speed
isoMDS_dat$speed <- isoMDS_dat$dist/1

fig4a <- ggplot(isoMDS_dat, aes(x, y, color = speed, shape = factor(cluster))) + geom_point() +
  theme_tufte(13) +
  # xlim(-0.10, .10) +
  # ylim(-0.12, 0.12) +
  scale_color_gradientn(colours=brewer.pal(7,"YlGnBu")) +
  annotate("text", x=0.12, y = 0.10, label='(A)', size = 5, color="black") +
  annotate("segment", x=-Inf, xend=Inf, y=-Inf, yend=-Inf, color = "grey") +
  annotate("segment", x=-Inf, xend=-Inf, y=-Inf, yend=Inf, color = "grey") +
  annotate("segment", x=Inf, xend=Inf, y=-Inf, yend=Inf, color = "grey") +
  annotate("segment", x=Inf, xend=-Inf, y=Inf, yend=Inf, color = "grey") +
  labs(x = "Dimension 1", 
       y= "Dimension 2", 
       color = "Speed", 
       shape = "Cluster") +
  theme(#aspect.ratio=1,
    legend.position = c(.5, .9),
    legend.box = "horizontal",
    legend.direction = 'horizontal',
    legend.justification = 'center',
    legend.title=element_text(size=8),
    legend.text = element_text(size=8),
    plot.title = element_text(hjust = 0.5)) +
  guides(color = guide_colorbar(order = 0, title.position = 'top', title.hjust = 0.5, barheight = .5),
         shape = guide_legend(order = 0, title.position = 'top', title.hjust = 0.5, barheight = .5)) +
  annotate("text", x = -0.09, y = -0.065, label = "k=2", size=4) +
  annotate("text", x = -0.09, y = -0.075, label = paste0("stress=", round(stress/100, 2)), size=4) +
  NULL
fig4a


fig4b <- ggplot(isoMDS_dat, aes(x=row, y=speed, color = factor(cluster))) + 
  geom_point(size=1.5, alpha = 0.7) +
  labs(x="Day of Jan-Feb", y="Speed of JS-Distance Divergence", color = "Cluster") +
  theme_tufte(13) +
  annotate("text", x=24*32, y = 0.21, label='(B)', size = 5, color="black") +
  geom_vline(xintercept = 14*24, color='grey') +
  geom_vline(xintercept = 20*24, color='grey') +
  annotate('text', x=17*24, y=0.205, label = "Event \n Window") +
  theme(legend.position = c(.10, .21),
        legend.box = "vertical",
        legend.box.background = element_rect(colour = "grey"),
        legend.direction = 'vertical',
        legend.justification = 'right',
        legend.title=element_text(size=8),
        legend.text = element_text(size=8),
        plot.title = element_text(hjust = 0.5)) +
  scale_x_continuous(breaks = c(1, 4*24, 9*24, 14*24, 16*24, 21*24, 26*24, 31*24), 
                     labels = c(16, 20, 25, 30, 1, 5, 10, 15)) +
  annotate("segment", x=-Inf, xend=Inf, y=-Inf, yend=-Inf, color = "grey") +
  annotate("segment", x=-Inf, xend=-Inf, y=-Inf, yend=Inf, color = "grey") +
  annotate("segment", x=Inf, xend=Inf, y=-Inf, yend=Inf, color = "grey") +
  annotate("segment", x=Inf, xend=-Inf, y=Inf, yend=Inf, color = "grey") +
  scale_color_manual(breaks = c("1", "2", "3"),
                     values = c("#440154FF", "#21908CFF", "darkorange"))
fig4b



# Horizontal align
# ggdraw() + draw_plot(fig4a, 0, 0, height = 1, width = .5) +
#   draw_plot(fig4b, .50, 0, height= 1, width = .5)

# Verticle align
ggdraw() + draw_plot(fig4a, 0, .50, height = .5, width = 1) + 
  draw_plot(fig4b, 0, 0, height= .5, width = 1)

  
ggsave("~/Projects/Anomalous-IUU-Events-Argentina/figures/figure_s2.pdf", width = 8, height = 8)
ggsave("~/Projects/Anomalous-IUU-Events-Argentina/figures/figure_s2.png", width = 8, height = 8)
  
  
#--------------------------------------------------------------------------------------------
# Figure S3 : Distance (log transformation))

# (A)

dat <- as.data.frame(read_feather('~/Projects/Anomalous-IUU-Events-Argentina/data/Argentina_5NN_region1_2018-02-05_2018-03-10.feather'))
dat <- dat %>% 
  mutate(day = day(timestamp),
         hour = hour(timestamp),
         month = month(timestamp)) %>% 
  filter(distance != 0) %>% 
  # filter(timestamp >= "2018-02-05 00:00:00" & timestamp <= "2018-03-06 23:00:00") %>% 
  group_by(timestamp, vessel_A) %>% 
  summarise(distance = mean(distance),
            ln_distance = mean(log(1 + distance)))

dat$date <- paste0(year(dat$timestamp), "-", month(dat$timestamp), "-", day(dat$timestamp), "-", hour(dat$timestamp))

cdat <- cut(dat$ln_distance, breaks = 50)

dat$breaks <- cdat 
dat$bin <- as.numeric(dat$breaks)


char <- unique(cdat)
retdat <- data.frame()
for (j in char){
  ldat <- data.frame(a = as.numeric(strsplit(gsub("\\[|\\]|\\(|\\)", "", j), ",")[[1]][1]),
                     b = as.numeric(strsplit(gsub("\\[|\\]|\\(|\\)", "", j), ",")[[1]][2]))
  retdat <- rbind(retdat, ldat)
}

outdat <- dat %>% 
  group_by(timestamp) %>% 
  mutate(nvessels = n()) %>% 
  group_by(timestamp, bin, nvessels) %>% 
  summarise(nbin_vessels = n()) %>% 
  mutate(prob = nbin_vessels/nvessels) %>% 
  ungroup()

outdat %>% 
  group_by(timestamp) %>% 
  summarise(sum = sum(prob))

outdat$day <- day(outdat$timestamp)


outdat$timestamp
as.Date.POSIXct(outdat$timestamp, c("%Y-%m-%d %h:00:00 PST"))
outdat$timestamp <- as.POSIXct(outdat$timestamp)

sb <- c(seq(5, 45, 5))
# Get log feet to dispaly on left side
ddat <- dat %>% 
  group_by(bin) %>% 
  summarise(m_dist = mean(ln_distance)) %>% 
  filter(bin %in% sb)
ddat
# ddat$x <- as.POSIXct("2016-03-09 23:00:00")

outdat <- filter(outdat, bin >= 5 & bin <= 45)

fig2a <- ggplot(outdat, aes(x=timestamp, y=factor(bin))) + 
  geom_tile(aes(fill = prob)) +
  theme_tufte(14) +
  # geom_vline(xintercept = 312, color ='red') +
  annotate("text", x=as.POSIXct("2018-03-9 2:00:00"), y = 41.5, label='(A)', size = 5, color  = "white") +
  labs(x="Day in Feb-March", y="Binned Distance(log km)", fill="P(d)") +
  scale_x_datetime(date_breaks = "2 day",
                   date_labels = "%d",
                   labels = c(seq(10, 21, 1)),
                   breaks = seq(10, 21, 1),
                   expand = expand_scale(mult = c(0, 0))) +
  scale_y_discrete(breaks = c(5, 10, 15, 20, 25, 30, 35, 40, 45), 
                   labels = round(ddat$m_dist, 1), 
                   expand = c(0, 0)) +
  scale_fill_gradientn(colours=rev(brewer.pal(11, "Spectral")), na.value = 'salmon') +
  annotate("rect", xmin = as.POSIXct("2018-02-19 01:00:00"),
           xmax = as.POSIXct("2018-02-25 01:00:00"),
           ymin = 0,
           ymax = 43,
           colour="white", alpha=0.1) +
  annotate("text", x = as.POSIXct("2018-02-22 00:00:00"), y=40, label="Event \n Window", color='white') +
  theme(legend.position = 'right',
        legend.margin=margin(l = 0, unit='cm'),
        panel.border = element_rect(colour = "grey", fill=NA, size=1),
        panel.grid = element_blank(),
        panel.background=element_rect(fill="#5E4FA2", colour="#5E4FA2")
  ) +
  guides(fill = guide_colorbar(label.hjust = unit(0, 'cm'),
                               frame.colour = "black",
                               barwidth = .5,
                               barheight = 10)) +
  # coord_fixed(ratio = 0.7) +
  # geom_label(data = ddat, aes(x=x, y=bin, label = round(m_dist, 1)), size = 2.5, nudge_y = -4) +
  NULL

fig2a
  
# (B)

# Figure 2 (B) 
dat <- as.data.frame(read_feather("~/Projects/Anomalous-IUU-Events-Argentina/data/dmat_Puerto_Madryn_region1_NN1_day_2018-02-05_2018-03-10.feather"))

dat$day <- seq(1, nrow(dat), 1)
pdat <- gather(dat, key = day, value = value)

pdat$day <- as.numeric(pdat$day) + 1
pdat$day2 <- seq(1, length(unique(pdat$day)))

fig2b <- ggplot(pdat, aes(x=day, y=day2)) + 
  theme_tufte(14) + 
  geom_tile(aes(fill=value)) +
  labs(y="Day in Feb-March", x="Day in Feb-March", fill="JS \nMetric") +
  # scale_fill_gradient(low='white', high='red') +
  scale_fill_gradientn(colours=rev(brewer.pal(11, "Spectral"))) +
  scale_y_continuous(trans = "reverse", breaks = c(1, 5, 10, 15, 20, 24, 29, 34),
                     labels = c(5, 10, 15, 20, 25, 1, 5, 10), expand=expand_scale(mult = c(0, 0))) +
  scale_x_continuous(breaks = c(1, 5, 10, 15, 20, 24, 29, 34),
                     labels = c(5, 10, 15, 20, 25, 1, 5, 10), expand=expand_scale(mult = c(0, 0))) +
  annotate("segment", x=-Inf, xend=Inf, y=-Inf, yend=-Inf, color = "black") +
  annotate("segment", x=-Inf, xend=-Inf, y=-Inf, yend=Inf, color = "black") +
  annotate("segment", x=Inf, xend=Inf, y=-Inf, yend=Inf, color = "black") +
  annotate("segment", x=Inf, xend=-Inf, y=Inf, yend=Inf, color = "black") +
  
  # Cluster 1
  annotate("segment", x=10, xend=10, y=0.5, yend=10, color = "black") +
  annotate("segment", x=0.5, xend=10, y=10, yend=10, color = "black") +
  # 
  # # Cluster 2
  annotate("segment", x=10, xend=22, y=10, yend=10, color = "black") +
  annotate("segment", x=10, xend=10, y=10, yend=22, color = "black") +
  annotate("segment", x=22, xend=22, y=10, yend=22, color = "black") +
  annotate("segment", x=10, xend=22, y=22, yend=22, color = "black") +
  # #
  # # Cluster 3
  annotate("segment", x=22, xend=22, y=22, yend=34.5, color = "black") +
  annotate("segment", x=22, xend=34.5, y=22, yend=22, color = "black") +
  # 
  annotate("text", x=7, y = 1, label='Cluster 1', size = 3) +
  annotate("text", x=19, y = 11, label='Cluster 2', size = 3) +
  annotate("text", x=31, y = 23, label='Cluster 3', size = 3) +
  annotate("text", x=33, y = 1.5, label='(B)', size = 5, color="white") +
  guides(fill = guide_colorbar(label.hjust = unit(0, 'cm'),
                               frame.colour = "black",
                               barwidth = .5,
                               barheight = 12)) +
  coord_equal() +
  NULL
fig2b


ggdraw() + draw_plot(fig2a, 0, .50, height = .50, width = 1) +
  draw_plot(fig2b, 0.01, -.25, height=1, width = 1)

ggsave("~/Projects/Anomalous-IUU-Events-Argentina/figures/figure_s3.png", width = 5, height = 8)
ggsave("~/Projects/Anomalous-IUU-Events-Argentina/figures/figure_s3.pdf", width = 5, height = 8)


#--------------------------------------------------------------------------------------------
# Figure S4 - time-series of the trailing rate of change in JS divergence as an index of behavioral change
  
dat <- as.data.frame(read_feather("~/Projects/Anomalous-IUU-Events-Argentina/data/dmat_Puerto_Madryn_region1_NN5_day-hour_2018-02-05_2018-03-10.feather"))


d <- as.matrix(dat)
fit <- isoMDS(d, k=2)
stress <- fit$stress
isoMDS_dat <- data.frame(x = fit$points[, 1], y = fit$points[, 2])


# Hourly cluster
#Clusters: 

# NN = 1 
cluster1 <- c(164, 0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 18, 19, 20, 21, 22, 23, 24, 25, 26, 27, 28, 29, 30, 31, 32, 33, 34, 35, 36, 37, 38, 39, 40, 41, 42, 43, 44, 45, 46, 47, 48, 49, 50, 51, 52, 53, 54, 55, 56, 57, 58, 59, 60, 61, 62, 63, 64, 65, 66, 67, 68, 69, 70, 71, 72, 73, 74, 75, 76, 77, 78, 79, 80, 81, 82, 83, 84, 85, 86, 87, 88, 89, 90, 91, 92, 93, 94, 95, 96, 97, 98, 99, 100, 101, 102, 103, 104, 105, 106, 107, 108, 109, 110, 111, 112, 113, 114, 115, 116, 117, 118, 119, 120, 121, 122, 123, 124, 125, 126, 127, 128, 129, 130, 131, 132, 133, 134, 135, 136, 137, 138, 139, 140, 141, 142, 143, 144, 145, 146, 147, 148, 149, 150, 151, 152, 153, 154, 155, 156, 157, 158, 159, 160, 161, 162, 163, 165, 166, 167, 168, 169, 170, 171, 172, 173, 174, 175, 176, 177, 178, 179, 180, 181, 182, 183, 184, 185, 186, 187, 188, 189, 190, 191, 192, 193, 194, 195, 196, 197, 198, 199, 200, 201, 202, 203, 204, 205, 206, 207, 208, 209, 210, 211, 212, 213, 214, 215, 216, 217, 218, 219, 220, 221, 222, 223, 224, 225, 226, 227, 228, 229, 230, 231, 232, 233, 234, 235, 236, 237, 238, 239, 240, 241, 242, 243, 244, 245, 246, 247, 248, 249, 250, 251, 252, 253, 254, 255, 256, 257, 258, 259, 260, 261, 262, 263, 264, 265, 266, 267, 268, 269, 270, 271, 272, 273, 274, 275, 276, 277, 278, 279, 280, 281, 282, 283, 284, 285, 286, 287, 288, 289, 290, 291, 292, 293, 294, 295, 296, 297, 298, 299, 300, 301, 302, 303, 304, 305, 306, 307, 308, 309, 310, 311, 312, 313, 314, 315, 316, 317, 318, 319, 320, 321, 322, 323, 324, 325, 327, 329, 330, 331, 332, 333, 334, 335, 336, 337, 338)
cluster2 <- c(460, 326, 328, 339, 340, 341, 342, 343, 344, 345, 346, 347, 348, 349, 350, 351, 352, 353, 354, 355, 356, 357, 358, 359, 360, 361, 362, 363, 364, 365, 366, 367, 368, 369, 370, 371, 372, 373, 374, 375, 376, 377, 378, 379, 380, 381, 382, 383, 384, 385, 386, 387, 388, 389, 390, 391, 392, 393, 394, 395, 396, 397, 398, 399, 400, 401, 402, 403, 404, 405, 406, 407, 408, 409, 410, 411, 412, 413, 414, 415, 416, 417, 418, 419, 420, 421, 422, 423, 424, 425, 426, 427, 428, 429, 430, 431, 432, 433, 434, 435, 436, 437, 438, 439, 440, 441, 442, 443, 444, 445, 446, 447, 448, 449, 450, 451, 452, 453, 454, 455, 456, 457, 458, 459, 461, 462, 463, 464, 465, 466, 467, 468, 469, 470, 471, 472, 473, 474, 475, 476, 477, 478, 479, 480, 481, 482, 483, 484, 485, 486, 487, 488, 489, 490, 491, 492, 493, 494, 495, 496, 497, 498, 499, 500, 501, 502, 503, 504, 505, 506, 507, 508, 509, 510, 511, 512, 513, 514, 515, 516, 517, 518, 519, 520, 521, 522, 523, 524, 525, 526, 527, 528, 529, 530, 531, 532, 533, 534, 535, 536, 537, 538, 539, 540, 541, 542, 543, 544, 545, 546, 547, 548, 549, 550, 551, 552, 553, 554, 555, 556, 557, 558, 559, 560, 561, 562, 563, 564, 565, 566, 567, 568, 569, 570, 571, 572, 573, 574, 575, 576, 577, 578, 579, 580, 581, 582, 583, 584, 585, 586, 587, 588, 589, 590, 591, 592, 593, 594, 595, 596, 597, 601, 603, 605, 606, 607, 608, 612, 613, 614, 615)
cluster3 <- c(723, 598, 599, 600, 602, 604, 609, 610, 611, 616, 617, 618, 619, 620, 621, 622, 623, 624, 625, 626, 627, 628, 629, 630, 631, 632, 633, 634, 635, 636, 637, 638, 639, 640, 641, 642, 643, 644, 645, 646, 647, 648, 649, 650, 651, 652, 653, 654, 655, 656, 657, 658, 659, 660, 661, 662, 663, 664, 665, 666, 667, 668, 669, 670, 671, 672, 673, 674, 675, 676, 677, 678, 679, 680, 681, 682, 683, 684, 685, 686, 687, 688, 689, 690, 691, 692, 693, 694, 695, 696, 697, 698, 699, 700, 701, 702, 703, 704, 705, 706, 707, 708, 709, 710, 711, 712, 713, 714, 715, 716, 717, 718, 719, 720, 721, 722, 724, 725, 726, 727, 728, 729, 730, 731, 732, 733, 734, 735, 736, 737, 738, 739, 740, 741, 742, 743, 744, 745, 746, 747, 748, 749, 750, 751, 752, 753, 754, 755, 756, 757, 758, 759, 760, 761, 762, 763, 764, 765, 766, 767, 768, 769, 770, 771, 772, 773, 774, 775, 776, 777, 778, 779, 780, 781, 782, 783, 784, 785, 786, 787, 788, 789, 790, 791, 792, 793, 794, 795, 796, 797, 798, 799, 800, 801, 802, 803, 804, 805, 806, 807, 808, 809, 810, 811, 812, 813, 814, 815, 816, 817, 818, 819, 820, 821, 822, 823, 824, 825, 826, 827, 828, 829, 830, 831, 832, 833, 834, 835, 836, 837, 838, 839, 840, 841, 842, 843, 844, 845, 846, 847, 848, 849, 850, 851, 852, 853, 854, 855, 856, 857, 858, 859, 860, 861, 862, 863, 864, 865, 866, 867, 868, 869, 870, 871, 872, 873, 874, 875, 876, 877, 878, 879, 880, 881, 882, 883, 884, 885, 886, 887, 888, 889, 890, 891, 892, 893, 894, 895, 896, 897, 898, 899, 900, 901, 902, 903, 904, 905, 906, 907, 908, 909, 910, 911)

#Medoids: [164, 460, 723]


# Hours
isoMDS_dat$row <- seq(1, nrow(isoMDS_dat))

# Merge data for cluster factors
clustdat <- data.frame(cluster = c(rep(1, length(cluster1)),
                                   rep(2, length(cluster2)),
                                   rep(3, length(cluster3))),
                       row = c(cluster1, cluster2, cluster3))

# Merge data
isoMDS_dat <- left_join(isoMDS_dat, clustdat, by = 'row')

# Calc distance t and t+1
# 2-axis
isoMDS_dat$dist <- sqrt( (isoMDS_dat$x - isoMDS_dat$y)^2 + (lead(isoMDS_dat$x) - lead(isoMDS_dat$y))^2 )

# Remove last obs
isoMDS_dat <- filter(isoMDS_dat, !is.na(dist))

# Calculate speed
isoMDS_dat$speed <- isoMDS_dat$dist/1

event_day <- filter(isoMDS_dat, row >= 17*24 & row <= 18*24)



# Color by speed (distance from day to lead(day))
fig4a <- ggplot(isoMDS_dat, aes(x, y, color = speed, shape = factor(cluster))) + geom_point() +
  # geom_point(data=event_day, aes(x, y), color='red') +
  annotate("text", x=0.12, y = 0.10, label='(A)', size = 5, color="black") +
  theme_tufte(13) +
  scale_color_gradientn(colours=brewer.pal(7,"YlGnBu")) +
  annotate("segment", x=-Inf, xend=Inf, y=-Inf, yend=-Inf, color = "grey") +
  annotate("segment", x=-Inf, xend=-Inf, y=-Inf, yend=Inf, color = "grey") +
  annotate("segment", x=Inf, xend=Inf, y=-Inf, yend=Inf, color = "grey") +
  annotate("segment", x=Inf, xend=-Inf, y=Inf, yend=Inf, color = "grey") +
  labs(x = "Dimension 1", 
       y= "Dimension 2", 
       color = "Speed", 
       shape = "Cluster") +
  theme(#aspect.ratio=1,
    legend.position = c(.5, .9),
    legend.box = "horizontal",
    legend.direction = 'horizontal',
    legend.justification = 'center',
    legend.title=element_text(size=8),
    legend.text = element_text(size=8),
    plot.title = element_text(hjust = 0.5)) +
  guides(shape = guide_legend(order = 1, title.position = 'top', title.hjust = 0.5, barheight = .5),
         color = guide_colorbar(order = 0, title.position = 'top', title.hjust = 0.5, barheight = .5)) +
  annotate("text", x = -0.09, y = -0.065, label = "k=2", size=4) +
  annotate("text", x = -0.09, y = -0.075, label = paste0("stress=", round(stress/100, 2)), size=4) +
  # coord_equal() +
  NULL

fig4a

fig4b <- ggplot(isoMDS_dat, aes(x=row, y=speed, color = factor(cluster))) + 
  geom_point(size=1.5, alpha = 0.7) +
  labs(x="Day of Feb-March", y="Speed of JS-Distance Divergence", color = "Cluster") +
  theme_tufte(13) +
  annotate("text", x=24*34, y = 0.23, label='(B)', size = 5, color="black") +
  geom_vline(xintercept = 13*24, color='grey') +
  geom_vline(xintercept = 19*24, color='grey') +
  annotate('text', x=16*24, y=0.225, label = "Event \n Window") +
  theme(legend.position = c(.10, .2),
        legend.box = "vertical",
        legend.box.background = element_rect(colour = "grey"),
        legend.direction = 'vertical',
        legend.justification = 'right',
        legend.title=element_text(size=8),
        legend.text = element_text(size=8),
        plot.title = element_text(hjust = 0.5)) +
  scale_x_continuous(breaks = c(1, 5*24-12, 10*24-12, 15*24-12, 20*24-12, 25*24-12, 29*24-12, 33*24-12), 
                     labels = c(5, 10, 15, 20, 25, 1, 5, 10)) +
  annotate("segment", x=-Inf, xend=Inf, y=-Inf, yend=-Inf, color = "grey") +
  annotate("segment", x=-Inf, xend=-Inf, y=-Inf, yend=Inf, color = "grey") +
  annotate("segment", x=Inf, xend=Inf, y=-Inf, yend=Inf, color = "grey") +
  annotate("segment", x=Inf, xend=-Inf, y=Inf, yend=Inf, color = "grey") +
  scale_color_manual(breaks = c("1", "2", "3"),
                     values = c("#440154FF", "#21908CFF", "#FFA500"))  

fig4b

# Horizontal align
# ggdraw() + draw_plot(fig4a, 0, 0, height = 1, width = .5) +
#   draw_plot(fig4b, .50, 0, height= 1, width = .5)

# Verticle align
dev.off()
ggdraw() + draw_plot(fig4a, 0, .50, height = .5, width = 1) + 
  draw_plot(fig4b, 0, 0, height= .5, width = 1)

ggsave("~/Projects/Anomalous-IUU-Events-Argentina/figures/figure_s4.png", width = 8, height = 8)
ggsave("~/Projects/Anomalous-IUU-Events-Argentina/figures/figure_s4.pdf", width = 8, height = 8)

# ------------------------------------------------
# Stress test
dat <- as.data.frame(read_feather("~/Projects/Anomalous-IUU-Events-Argentina/data/dmat_Puerto_Madryn_region1_NN1_day-hour_2016-03-01_2016-03-31.feather"))

d <- as.matrix(dat)
fit <- isoMDS(d, k=5)

isoMDS_dat <- data.frame(x1 = fit$points[, 1], x2 = fit$points[, 2], x3 = fit$points[, 3], x4 = fit$points[, 4], x5 = fit$points[, 5])

stress <- fit$stress
isoMDS_dat$dist <- sqrt(((isoMDS_dat$x1 - isoMDS_dat$x2 - isoMDS_dat$x3 - isoMDS_dat$x4 - isoMDS_dat$x5)^2 + (lead(isoMDS_dat$x1) - lead(isoMDS_dat$x2) - lead(isoMDS_dat$x3) - lead(isoMDS_dat$x4) - lead(isoMDS_dat$x5))^2))
isoMDS_dat$speed <- isoMDS_dat$dist/1
head(isoMDS_dat)

# Remove last obs
isoMDS_dat <- filter(isoMDS_dat, !is.na(dist))

isoMDS_dat$row <- 1:nrow(isoMDS_dat)

# Color by speed (distance from day to lead(day))
fig3b <- ggplot(isoMDS_dat, aes(x=row, y=speed)) + 
  geom_point(size=1.5, alpha = 0.7) +
  labs(x="Day in March", y="Speed of JS-Distance Divergence", color = "Cluster") +
  # annotate("text", x=24*31, y = 0.245, label='(B)', size = 5, color="black") +
  theme_tufte(13) +
  geom_vline(xintercept = 12*24, color='grey') +
  geom_vline(xintercept = 18*24, color='grey') +
  annotate('text', x=15*24, y=0.32, label = "Event \n Window") +
  theme(legend.position = c(.10, .2),
        legend.box = "vertical",
        legend.box.background = element_rect(colour = "grey"),
        legend.direction = 'vertical',
        legend.justification = 'right',
        legend.title=element_text(size=8),
        legend.text = element_text(size=8),
        plot.title = element_text(hjust = 0.5)) +
  scale_x_continuous(breaks = c(1, 5*24, 10*24, 15*24, 20*24, 25*24, 31*24), labels = c(1, 5, 10, 15, 20, 25, 31)) +
  annotate("segment", x=-Inf, xend=Inf, y=-Inf, yend=-Inf, color = "grey") +
  annotate("segment", x=-Inf, xend=-Inf, y=-Inf, yend=Inf, color = "grey") +
  annotate("segment", x=Inf, xend=Inf, y=-Inf, yend=Inf, color = "grey") +
  annotate("segment", x=Inf, xend=-Inf, y=Inf, yend=Inf, color = "grey") +
  #scale_color_viridis(discrete=TRUE) +
  scale_color_manual(breaks = c("1", "2", "3"),
                     values = c("#440154FF", "#21908CFF", "darkorange")) +
  NULL

fig3b

# Horizontal align
# ggdraw() + draw_plot(fig3a, 0, 0, height = 1, width = .5) +
#  draw_plot(fig3b, .50, 0, height= 1, width = .5)

# Verticle align
ggdraw() + draw_plot(fig3a, 0, .50, height = .5, width = 1) + 
  draw_plot(fig3b, 0, 0, height= .5, width = 1)

# ggsave("~/Projects/Anomalous-IUU-Events-Argentina/figures/figure3.png", width = 8, height = 8)
# ggsave("~/Projects/Anomalous-IUU-Events-Argentina/figures/figure3.pdf", width = 8, height = 8)








#--------------------------------------------------------------------------------------------
# Figure S5 - JS-Distance t and t+1

dat <- as.data.frame(read_feather('~/Projects/Anomalous-IUU-Events-Argentina/data/Argentina_5NN_region1_2016-03-01_2016-03-31.feather'))
fig3 <- filter(dat, distance != 0)
fig3 <- filter(fig3, NN <= 5)
fig3 <- fig3 %>% 
  group_by(timestamp, vessel_A) %>% 
  summarise(distance = mean(distance))
fig3$ln_distance <- log(1 + fig3$distance)
fig3$month <- month(fig3$timestamp)
fig3$day <- day(fig3$timestamp)
fig3$hour <- hour(fig3$timestamp)




breaks=seq(min(data),max(data),l=number_of_bins+1)

x = 
bw <- 2 * IQR(q) / length(q)^(1/3)
num_bins <- diff(range(p)) / (2 * IQR(p) / length(p)^(1/3))
ph <- graphics::hist(p, breaks="FD", plot=FALSE)
qh <- graphics::hist(q, breaks=bw, plot=FALSE)
ph <- ph$counts/sum(ph$counts)
qh <- qh$counts/sum(qh$counts)

JSD(rbind(qh, ph))

jsd_dat <- data.frame()
for (i in unique(fig3$timestamp)){
  print(i)
  p <- filter(fig3, timestamp == i )
  q <- filter(fig3, timestamp == i + 60*60) # Plus 1 hour (60 seconds * 60 minutes)
  date = p$timestamp[1]
  p <- log(1 + p$distance)
  q <- log(1 + q$distance)
  num_bins <- diff(range(p)) / (2 * IQR(p) / length(p)^(1/3))
  ph <- hist(p, breaks=40, plot=FALSE)
  qh <- hist(q, breaks=40, plot=FALSE)
  pp <- ph$counts/sum(ph$counts)
  qq <- qh$counts/sum(qh$counts)
  jsd_d <- sqrt(JSD(rbind(pp, qq)))
  #for (j in 1:1000){
  #  pp <- sample(p, minobs, replace = TRUE)
  #  qq <- sample(q, minobs, replace = TRUE)
  #  js <- JSD(qq, pp)
  #  indat <- data.frame(js = sqrt(js))
  #  lst <- rbind(lst, indat)
  #}
  #outdat <- data.frame(t = date, jsd_mean = mean(lst$js), jsd_sd = sd(lst$js))
  outdat <- data.frame(t = date, jsd = jsd_d)
  jsd_dat <- rbind(jsd_dat, outdat)
}

plot(jsd_dat$jsd)

jsd_dat$day <- day(jsd_dat$t)
jsd_dat$hour <- hour(jsd_dat$t)
jsd_dat2 <- jsd_dat %>% 
  group_by(t) %>% 
  summarise(jsd_mean = mean(jsd_mean))

jsd_dat2$day <- seq(1, nrow(jsd_dat2), 1)

jsd_s5a <- jsd_dat2

s5a <-ggplot(jsd_s5a, aes(x=day, y=jsd_mean)) + 
  theme_tufte(13) +
  geom_point(alpha=0.7) +
  labs(x="Day in March 2016 ", y="Leading JS-Distance") +
  theme(panel.border = element_rect(colour = "grey", fill=NA, size=1)) +
  scale_x_continuous(breaks = c(1, 5*24, 10*24, 15*24, 20*24, 25*24, 31*24), labels = c(1, 5, 10, 15, 20, 25, 31)) +
  annotate("segment", x=-Inf, xend=Inf, y=-Inf, yend=-Inf, color = "grey") +
  annotate("segment", x=-Inf, xend=-Inf, y=-Inf, yend=Inf, color = "grey") +
  annotate("segment", x=Inf, xend=Inf, y=-Inf, yend=Inf, color = "grey") +
  annotate("segment", x=Inf, xend=-Inf, y=Inf, yend=Inf, color = "grey") +
  # ylim(0.33, 0.4) +
  scale_y_continuous(breaks = c(0.33, 0.34, 0.35, 0.36, 0.37, 0.38, 0.39, 0.40, 0.41), limits = c(0.33, 0.41)) +
  annotate("text", x=24*31-6, y = 0.41, label='(A)', size = 5, color="black") +
  NULL

s5a

dat <- as.data.frame(read_feather('~/Projects/Anomalous-IUU-Events-Argentina/data/Argentina_5NN_region1_2018-01-15_2018-02-15.feather'))
fig3 <- filter(dat, distance != 0)
fig3 <- filter(fig3, NN <= 5)
fig3 <- fig3 %>% 
  group_by(timestamp, vessel_A) %>% 
  summarise(distance = mean(distance))
fig3$ln_distance <- log(1 + fig3$distance)
fig3$month <- month(fig3$timestamp)
fig3$day <- day(fig3$timestamp)
fig3$hour <- hour(fig3$timestamp)

jsd_dat <- data.frame()
for (i in unique(fig3$timestamp)){
  print(i)
  p <- filter(fig3, timestamp == i )
  q <- filter(fig3, timestamp == i + 60*60) # Plus 1 hour (60 seconds * 60 minutes)
  date = p$timestamp[1]
  p <- p$distance
  q <- q$distance
  qmin <- length(q)
  pmin <- length(p)
  minobs <- min(qmin, pmin)
  lst <- data.frame()
  for (j in 1:1000){
    pp <- sample(p, minobs, replace = TRUE)
    qq <- sample(q, minobs, replace = TRUE)
    js <- JSD(qq, pp)
    indat <- data.frame(js = sqrt(js))
    lst <- rbind(lst, indat)
  }
  outdat <- data.frame(t = date, jsd_mean = mean(lst$js), jsd_sd = sd(lst$js))
  jsd_dat <- rbind(jsd_dat, outdat)
}

jsd_dat$day <- day(jsd_dat$t)
jsd_dat$hour <- hour(jsd_dat$t)
jsd_dat2 <- jsd_dat %>% 
  group_by(t) %>% 
  summarise(jsd_mean = mean(jsd_mean))

jsd_dat2$day <- seq(1, nrow(jsd_dat2), 1)

jsd_s5b <- jsd_dat2

s5b <-ggplot(jsd_s5b, aes(x=day, y=jsd_mean)) + 
  theme_tufte(13) +
  geom_point(alpha=0.7) +
  labs(x="Day in Jan-Feb 2018", y="Leading JS-Distance") +
  theme(panel.border = element_rect(colour = "grey", fill=NA, size=1)) +
  scale_x_continuous(breaks = c(1, 4*24, 9*24, 14*24, 16*24, 21*24, 26*24, 31*24), labels = c(16, 20, 25, 30, 1, 5, 10, 15)) +
  annotate("segment", x=-Inf, xend=Inf, y=-Inf, yend=-Inf, color = "grey") +
  annotate("segment", x=-Inf, xend=-Inf, y=-Inf, yend=Inf, color = "grey") +
  annotate("segment", x=Inf, xend=Inf, y=-Inf, yend=Inf, color = "grey") +
  annotate("segment", x=Inf, xend=-Inf, y=Inf, yend=Inf, color = "grey") +
  # ylim(0.33, 0.4) +
  scale_y_continuous(breaks = c(0.33, 0.34, 0.35, 0.36, 0.37, 0.38, 0.39, 0.40, 0.41), limits = c(0.33, 0.41)) +
  annotate("text", x=24*32-6, y = 0.41, label='(B)', size = 5, color="black") +
  NULL

s5b

dat <- as.data.frame(read_feather('~/Projects/Anomalous-IUU-Events-Argentina/data/Argentina_5NN_region1_2018-02-05_2018-03-10.feather'))
fig3 <- filter(dat, distance != 0)
fig3 <- filter(fig3, NN <= 5)
fig3 <- fig3 %>% 
  group_by(timestamp, vessel_A) %>% 
  summarise(distance = mean(distance))
fig3$ln_distance <- log(1 + fig3$distance)
fig3$month <- month(fig3$timestamp)
fig3$day <- day(fig3$timestamp)
fig3$hour <- hour(fig3$timestamp)

jsd_dat <- data.frame()
for (i in unique(fig3$timestamp)){
  print(i)
  p <- filter(fig3, timestamp == i )
  q <- filter(fig3, timestamp == i + 60*60) # Plus 1 hour (60 seconds * 60 minutes)
  date = p$timestamp[1]
  p <- p$distance
  q <- q$distance
  qmin <- length(q)
  pmin <- length(p)
  minobs <- min(qmin, pmin)
  lst <- data.frame()
  for (j in 1:1000){
    pp <- sample(p, minobs, replace = TRUE)
    qq <- sample(q, minobs, replace = TRUE)
    js <- JSD(qq, pp)
    indat <- data.frame(js = sqrt(js))
    lst <- rbind(lst, indat)
  }
  outdat <- data.frame(t = date, jsd_mean = mean(lst$js), jsd_sd = sd(lst$js))
  jsd_dat <- rbind(jsd_dat, outdat)
}

jsd_dat$day <- day(jsd_dat$t)
jsd_dat$hour <- hour(jsd_dat$t)
jsd_dat2 <- jsd_dat %>% 
  group_by(t) %>% 
  summarise(jsd_mean = mean(jsd_mean))

jsd_dat2$day <- seq(1, nrow(jsd_dat2), 1)

jsd_s5c <- jsd_dat2

s5c <-ggplot(jsd_s5c, aes(x=day, y=jsd_mean)) + 
  theme_tufte(13) +
  geom_point(alpha=0.7) +
  labs(x="Day in Feb-Mar 2018", y="Leading JS-Distance") +
  theme(panel.border = element_rect(colour = "grey", fill=NA, size=1)) +
  scale_x_continuous(breaks = c(1, 5*24-12, 10*24-12, 15*24-12, 20*24-12, 25*24-12, 29*24-12, 33*24-12), labels = c(5, 10, 15, 20, 25, 1, 5, 10)) +
  annotate("segment", x=-Inf, xend=Inf, y=-Inf, yend=-Inf, color = "grey") +
  annotate("segment", x=-Inf, xend=-Inf, y=-Inf, yend=Inf, color = "grey") +
  annotate("segment", x=Inf, xend=Inf, y=-Inf, yend=Inf, color = "grey") +
  annotate("segment", x=Inf, xend=-Inf, y=Inf, yend=Inf, color = "grey") +
  scale_y_continuous(breaks = c(0.33, 0.34, 0.35, 0.36, 0.37, 0.38, 0.39, 0.40, 0.41), limits = c(0.33, 0.41)) +
  annotate("text", x=24*34-6, y = 0.41, label='(C)', size = 5, color="black") +
  NULL

plot_grid(s5a, s5b, s5c)

ggsave("~/Projects/Anomalous-IUU-Events-Argentina/figures/figure_s5.png", width = 8, height = 6)
ggsave("~/Projects/Anomalous-IUU-Events-Argentina/figures/figure_s5.pdf", width = 8, height = 6)


#--------------------------------------------------------------
# Animated Plot
bat <- getNOAA.bathy(-68, -51, -48, -39, res = 1, keep = TRUE)
dat <- as.data.frame(read_feather('~/Projects/Anomalous-IUU-Events-Argentina/data/Argentina_5NN_region1_2016-03-01_2016-03-31.feather'))
dat$month <- month(dat$timestamp)
dat$day <- day(dat$timestamp)
dat$hour <- hour(dat$timestamp)
dat$year <- year(dat$timestamp)
dat$min <- minute(dat$timestamp)
dat$ln_distance <- log(1 + dat$distance)

dat$month <- stringr::str_pad(dat$month, 2, side = "left", pad = 0)
dat$day <- stringr::str_pad(dat$day, 2, side = "left", pad = 0)
dat$hour <- stringr::str_pad(dat$hour, 2, side = "left", pad = 0)
dat$min <- stringr::str_pad(dat$min, 2, side = "left", pad = 0)

movdat <- dat

NN_max <- 3

movdat <- filter(movdat, NN <= NN_max)
movdat <- filter(movdat, month == "03")

#i = movdat$timestamp[100000]
i = "2016-03-15 12:00:00"
i = "2016-03-15 19:00:00"
i = "2016-03-29 00:00:00"

for (i in unique(movdat$timestamp)){
  
  subdat <- filter(movdat, timestamp == i)
  
  date_ <- paste0(subdat$year[1], "-", subdat$month[1], "-", subdat$day[1], " ", subdat$hour[1], ":", subdat$min[1], ":00")
  date_
  
  #  #C6E0FC
  
  movmap <- 
    # ggplot(bat, aes(x, y, z)) +
    autoplot(bat, geom = c("raster", "contour")) +
    geom_raster(aes(fill=z)) +
    geom_contour(aes(z = z), colour = "white", alpha = 0.05) +
    scale_fill_gradientn(values = scales::rescale(c(-6600, 0, 1, 1500)),
                         colors = c("lightsteelblue4", "lightsteelblue2", "#C6E0FC", 
                                    "grey50", "grey70", "grey85")) +
    labs(x=NULL, y=NULL, color="km") +
    geom_segment(data=subdat, aes(x=vessel_A_lon, xend=vessel_B_lon, y=vessel_A_lat, yend=vessel_B_lat, color=distance), size = 0.1) +
    geom_segment(data=subdat, aes(x=vessel_B_lon, xend=vessel_A_lon, y=vessel_B_lat, yend=vessel_A_lat, color=distance), size = 0.1) +
    geom_point(data=subdat, aes(x=vessel_A_lon, y=vessel_A_lat), size = .5, color='red', alpha=0.7) +
    annotate("text", x=-53.75, y = -39.25, label=date_, size = 4, color='black', fontface=2) +
    theme(axis.title.x=element_blank(),
          axis.text.x=element_blank(),
          axis.ticks.x=element_blank(),
          axis.title.y=element_blank(),
          axis.text.y=element_blank(),
          axis.ticks.y=element_blank(),
          legend.direction = 'vertical',
          legend.justification = 'center',
          legend.position = c(.95, 0.24),
          legend.margin=margin(l = 0, unit='cm'),
          legend.text = element_text(size=10),
          legend.title = element_text(size=12),
          panel.grid = element_blank()) +
    scale_color_gradientn(colours=brewer.pal(9, "YlOrRd"), limits=c(0, 200)) +
    
    # Legend up top
    guides(fill = FALSE,
           color = guide_colorbar(title.hjust = unit(1.1, 'cm'),
                                  title.position = "top",
                                  frame.colour = "black",
                                  barwidth = .5,
                                  barheight = 7,
                                  label.position = 'left')) +
    annotate("segment", x=-Inf, xend=Inf, y=-Inf, yend=-Inf, color = "black", size=1) + #Bottom
    annotate("segment", x=-Inf, xend=-Inf, y=-Inf, yend=Inf, color = "black", size=1) + # Left
    annotate("segment", x=Inf, xend=Inf, y=-Inf, yend=Inf, color = "black", size=1) + # Right
    annotate("segment", x=-Inf, xend=Inf, y=Inf, yend=Inf, color = "black", size=1) + # Top
    # scale_y_continuous(expand=c(0,0)) +
    # scale_x_continuous(expand=c(0,0)) +
    NULL
  movmap
  
  
  ggsave(filename = paste0("~/Projects/Anomalous-IUU-Events-Argentina/figures/animated_fig/hourly_figs/", date_, ".png"), width = 6, height = 4, plot = movmap)
}
